Move Quality into public.h
