+++ samples/include/samples.h
@@ -14,7 +14,6 @@
 
 #pragma once
 #include "manifold.h"
-#include "quality.h"
 
 namespace manifold {
